requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
haversine==2.9.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
pillow_heif==0.22.0
pluggy==1.5.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.39
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
